pyproject.toml
[tool.poetry]
name = "zkbridge"
version = "0.1.0"
description = ""
authors = ["Alen <alen.kimov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"
numpy = "^1.24.3"
aiohttp = "^3.8.4"
requests = "^2.31.0"
loguru = "^0.7.0"
web3 = "^6.4.0"
pydantic = "^1.10.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
